requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.8.2